backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
python-multipart==0.0.6
cryptography==41.0.7
gspread==5.12.0
google-auth==2.23.4
watchdog==3.0.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
# psycopg2 powers the sync engine used by PgStore (dict-facade from router sync contexts)
# and by conftest.py for TRUNCATE-based test resets.
psycopg2-binary==2.9.10
pgvector==0.3.6
alembic==1.14.0
greenlet==3.1.1

# Dev / testing
pytest==8.1.1
pytest-asyncio==0.23.6
